Moved ComparisonPlotter class to separate file
